+++ src/main/kotlin/mobi/hsz/idea/gitignore/settings/IgnoreSettings.kt
@@ -24,7 +24,7 @@ class IgnoreSettings : PersistentStateComponent<Element?>, Listenable<IgnoreSett
         USER_TEMPLATES_TEMPLATE("template"), USER_TEMPLATES_NAME("name"), LANGUAGES("languages"),
         LANGUAGES_LANGUAGE("language"), LANGUAGES_ID("id"), IGNORED_FILE_STATUS("ignoredFileStatus"),
         INSERT_AT_CURSOR("insertAtCursor"), STARRED_TEMPLATES("starredTemplates"), UNIGNORE_ACTIONS("unignoreActions"),
-        NOTIFY_IGNORED_EDITING("notifyIgnoredEditing"), FOLDABLE_PROJECT_VIEW_ADVERTISE_PRESENTED("foldableProjectViewAdvertisePresented");
+        NOTIFY_IGNORED_EDITING("notifyIgnoredEditing");
 
         override fun toString() = key
     }
@@ -150,7 +150,6 @@ class IgnoreSettings : PersistentStateComponent<Element?>, Listenable<IgnoreSett
         setAttribute(KEY.STARRED_TEMPLATES.toString(), StringUtil.join(starredTemplates, Constants.DOLLAR))
         setAttribute(KEY.UNIGNORE_ACTIONS.toString(), unignoreActions.toString())
         setAttribute(KEY.NOTIFY_IGNORED_EDITING.toString(), notifyIgnoredEditing.toString())
-        setAttribute(KEY.FOLDABLE_PROJECT_VIEW_ADVERTISE_PRESENTED.toString(), foldableProjectViewAdvertisePresented.toString())
 
         addContent(
             Element(KEY.LANGUAGES.toString()).apply {
@@ -185,9 +184,6 @@ class IgnoreSettings : PersistentStateComponent<Element?>, Listenable<IgnoreSett
             }
             getAttributeValue(KEY.NOTIFY_IGNORED_EDITING.toString())?.let {
                 notifyIgnoredEditing = it.toBoolean()
-            }
-            getAttributeValue(KEY.FOLDABLE_PROJECT_VIEW_ADVERTISE_PRESENTED.toString())?.let {
-                foldableProjectViewAdvertisePresented = it.toBoolean()
             }
 
             getChild(KEY.LANGUAGES.toString()).children.forEach {